@@ -6,14 +6,14 @@ class ConvertEditPermissionsToUpdate < ActiveRecord::Migration
   def up
     Admin::Permission.reset_column_information
     Admin::Permission.where(:action => "edit").each do |product|
-      product.update_attributes!(:action => "update")
+      product.update!(:action => "update")
     end
   end
 
   def down
     Admin::Permission.reset_column_information
     Admin::Permission.where(:action => "update").each do |product|
-      product.update_attributes!(:action => "edit")
+      product.update!(:action => "edit")
     end
   end
 end
