@@ -18,18 +18,18 @@ class Admin::PermissionsController < AdminController
   # Takes the data posted from the grid and sets the permissions.
   ##
   def update_grid
-    @roles.each do |role|
+    @roles.includes(:permissions).each do |role|
       models = params[role.name]
 
       (@models.map(&:name) + @miscellaneous_permission_subject_classes.keys).uniq.each do |model_name|
-        actions = models[model_name] if models
+        actions = models[model_name]&.keys if models
         actions ||= []
 
         Admin::Permission.update_permission(role, model_name, actions)
       end
     end
 
-    return redirect_to :admin_permissions
+    redirect_to admin_permissions_url
   end
 
   private
@@ -43,7 +43,7 @@ class Admin::PermissionsController < AdminController
       'Event' => { 'add_non_members' => 'Add non-members to events, mainly for archiving purposes' },
     }
 
-    @models = ApplicationRecord.descendants + [Admin::Debt, Season]
+    @models = (ApplicationRecord.descendants + [Admin::Debt, Season]).uniq
 
     role_exclude = ['admin', 'Proposal Checker']
     @roles = Role.includes(:permissions).where.not(name: role_exclude).all
