@@ -14,7 +14,7 @@
 #++
 ##
 class Admin::Questionnaires::QuestionnaireTemplate < ApplicationRecord
-  validates :name, presence: true, uniqueness: true
+  validates :name, presence: true, uniqueness: { case_sensitive: false }
 
   has_many :questions, as: :questionable
 
