@@ -28,5 +28,5 @@
 class Workshop < Event
   # Validate uniqueness on Event Subtype basis instead of on the event.
   # Otherwise, you cannot have two different types with the same slug.
-  validates :slug, uniqueness: true
+  validates :slug, uniqueness: { case_sensitive: false }
 end
